task - subtask relation

+++ src/main/kotlin/com/piffek/todoapp/persistance/task/TaskRepository.kt
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TaskRepository : JpaRepository<Task, Int> {}
+interface TaskRepository : JpaRepository<Task, Long> {}
